feat(queries): get it working, with tests, with refactoring
