--- config/gql-to-app-runtime-query/src/__test__/plugin/queryFailureWrongDefaultVariablesGraphql.js
@@ -1,9 +1,0 @@
-import { gql } from '@dhis2/app-runtime'
-
-const query = gql`
-    query queryName($id: String = 42) {
-        SmsCommand(resource: "smsCommands", id: $id) {
-            __all
-        }
-    }
-`
