+++ config/gql-to-app-runtime-query/src/__test__/plugin/success-queries/default-values.transformed.js
@@ -3,10 +3,40 @@ const query = {
   SmsCommand: {
     resource: "smsCommands",
     params: ({
-      test = "test"
+      str = "test",
+      int = 42,
+      float = 4.2,
+      boolTrue = true,
+      boolFalse = false,
+      arr = ["test", 42, 4.2, true, false, {
+        foo: "foo"
+      }, [{
+        bar: "bar"
+      }]],
+      obj = {
+        str: "test",
+        int: 42,
+        float: 4.2,
+        boolTrue: true,
+        boolFalse: false,
+        foo: ["test", 42, 4.2, true, false, {
+          foo: "foo"
+        }, [{
+          bar: "bar"
+        }]],
+        bar: {
+          baz: "baz"
+        }
+      }
     }) => ({
       fields: ["*"],
-      test: test
+      str: str,
+      int: int,
+      arr: arr,
+      obj: obj,
+      boolTrue: boolTrue,
+      boolFalse: boolFalse,
+      float: float
     })
   }
 };
