+++ config/gql-to-app-runtime-query/src/__test__/plugin/success-queries/multiple-resources.transformed.js
@@ -1,0 +1,15 @@
+import { gql } from '@dhis2/app-runtime';
+const query = {
+  SmsCommand: {
+    resource: "smsCommands",
+    params: ({}) => ({
+      fields: ["id", "displayName", "dataElements[id,displayName,dataSets[id,displayName]]"]
+    })
+  },
+  DataElement: {
+    resource: "dataElements",
+    params: ({}) => ({
+      fields: ["id", "displayName", "dataSetElements[categoryCombo[*]]"]
+    })
+  }
+};
