+++ config/gql-to-app-runtime-query/src/__test__/plugin/success-queries/nested-fields.transformed.js
@@ -3,7 +3,7 @@ const query = {
   SmsCommand: {
     resource: "smsCommands",
     params: ({}) => ({
-      fields: ["id", "displayName", "dataElements[id,displayName,dataSets[id,displayName]]"]
+      fields: ["id", "displayName", "dataElements[id,displayName,dataSets[id,displayName]]", "smsCodes[code,optionId]"]
     })
   }
 };
