+++ config/gql-to-app-runtime-query/src/handleQuery/createRuntimeResource/getFirstLevelFields.js
@@ -1,0 +1,13 @@
+const { pipe } = require('./pipe')
+
+const transformSpecialFields = field => {
+    if (field === '__all') return '*'
+    return field
+}
+
+const getFirstLevelFields = pipe(
+    curFields => curFields.filter(field => !(field instanceof Object)),
+    curFields => curFields.map(transformSpecialFields)
+)
+
+module.exports.getFirstLevelFields = getFirstLevelFields
