+++ config/gql-to-app-runtime-query/src/handleQuery/createRuntimeResource/getNestedLevelFields.js
@@ -1,0 +1,4 @@
+const getNestedLevelFields = curFields =>
+    curFields.filter(field => field instanceof Object)
+
+module.exports.getNestedLevelFields = getNestedLevelFields
