+++ config/gql-to-app-runtime-query/src/handleQuery/createRuntimeResource/index.js
@@ -1,0 +1,2 @@
+const { createRuntimeResource } = require('./createRuntimeResource')
+module.exports.createRuntimeResource = createRuntimeResource
