+++ config/gql-to-app-runtime-query/src/handleQuery/createRuntimeResource/pipe.js
@@ -1,0 +1,6 @@
+const pipe = (...fns) => (...args) => {
+    if (!fns.length) return
+    return fns.slice(1).reduce((acc, fn) => fn(acc), fns[0](...args))
+}
+
+module.exports.pipe = pipe
