+++ config/gql-to-app-runtime-query/src/handleQuery/extractResources/index.js
