+++ config/gql-to-app-runtime-query/src/handleQuery/extractVariables/getTypeOfVariable.js
