+++ config/gql-to-app-runtime-query/src/handleQuery/extractVariables/validateVariableDefaultValues.js
