+++ config/gql-to-app-runtime-query/src/plugin/index.js
@@ -1,0 +1,2 @@
+const { plugin } = require('./plugin')
+module.exports.plugin = plugin
