+++ config/gql-to-app-runtime-query/src/plugin/validateIsGqlTemplateExpression/index.js
