--- config/gql-to-app-runtime-query/src/resourceToObjectProperty/index.js
@@ -1,2 +1,0 @@
-const { resourceToObjectProperty } = require('./resourceToObjectProperty')
-module.exports.resourceToObjectProperty = resourceToObjectProperty
