--- config/gql-to-app-runtime-query/src/validateVariableDefaultValues/index.js
@@ -1,4 +1,0 @@
-const {
-    validateVariableDefaultValues,
-} = require('./validateVariableDefaultValues')
-module.exports.validateVariableDefaultValues = validateVariableDefaultValues
